Added `priority` fields to `action_on_being_used` and `action_on_entity_use`
The powers now trigger in order of descending priority, and interactions with lower priority are only triggered when all interactions with higher priority have `pass` as the `action_result`.
The vanilla entity use interaction is considered to be priority 0.
The default `priority` for the powers is 0.
Existing datapacks will only change in behaviour if they had `pass` as the action result.