Created Configurable Parameters for locateBiome()
This adds 3 configurable fields to the config.

Radius: The radius of the search.

HorizontalBlockCheckInterval: The number of blocks between each check in the x and z direction.

VerticalBlockCheckInterval: The number of blocks between each check in the y direction.

These values are by default 6400, 64, and 64 respectively. Setting the radius to 0 would essentially disable the modifyPlayerSpawnPower. Setting any of these values to a negative number will set them back to their default values. The same is applied whenever horizontalBlockCheckInterval and verticalBlockCheckInterval to 0