Added a parent/child system to power types.
adds an optional "parent" field to power types that accepts a power id.
the JsonObjects of the powers, will then be merged. with the child's fields taking priority over the parents, allowing for overriding actions or conditions.

NOTE: You cannot make a child of a child, because the parent field will never be overwriten by that of a child. so the sub-child will never have a reference to the original.