Memory leak & threading fixes (#260)
* Fix a possible race condition in EntityTypeMixin

Client & server both iterate and modify the EntityType, which can be a possible issue with a singleplayer world.

* Fix memory leak via EntityType

* Fix memory leak via permanently retained ItemStack

* Fix potential thread safety issue in DefaultAttributeContainerMixin

Both the client and server could try to modify these at the same time in the case of singleplayer.

* Update Actions

* Actually initialise the ThreadLocal

* Fix memory leak with DefaultAttributeContainer

Refactored the internal calls to not be easily discoverable with a blatant warning that it's not supposed to be public API.

* Fix another memory leak with DefaultAttributeContainer

Evidently, the EntityAttributeInstances that were present also have owner entities attached.

* Fix possible memory leak with ItemStack

As it is not possible to properly fix this without breaking API, WeakReference was used.

* This one seems to be insistent on leaking and I frankly don't understand why.

* Fix off-thread being able to also leak entities.

* Don't strongly retain players in RecipeBook

This is primarily only relevant to the client, where
it has been observed that the player retained a player,
causing double the memory leak, especially when the
world was also leaked.